Refactored code, moved IconLogo to Component
